saved json files in the architecture folder
