+++ scripts/createOutput.sh
@@ -24,4 +24,4 @@ node scripts/fixImages.js target/generated-docs/solutions
 
 node scripts/createSolutions.js target/generated-docs/solutions target/generated-docs/solutions.json 500
 node scripts/createTags.js solutions target/generated-docs/tags.json
-node scripts/createIndex.js target/generated-docs/solutions target/generated-docs/index.json
+node scripts/createIndex.js target/generated-docs/solutions .html target/generated-docs/docs-json.json target/generated-docs/index.json
